\name{TreeFactor-package}
\alias{TreeFactor-package}
\alias{TreeFactor}
\docType{package}
\title{
\packageTitle{TreeFactor}
}
\description{
\packageDescription{TreeFactor}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{TreeFactor}
\packageIndices{TreeFactor}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{TreeFactor}

Maintainer: \packageMaintainer{TreeFactor}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
